Relocate geoServer api
